+++ tests/src/test_memfault_ram_reboot_tracking.cpp
@@ -18,6 +18,7 @@ TEST_GROUP(MfltStorageTestGroup) {
     memset(&s_mflt_reboot_tracking_region[0], 0xBA, sizeof(s_mflt_reboot_tracking_region));
     memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, NULL);
     memfault_reboot_tracking_reset_crash_count();
+    memfault_reboot_tracking_clear_reboot_reason();
     memfault_reboot_tracking_clear_reset_info();
   }
 
@@ -75,14 +76,12 @@ TEST(MfltStorageTestGroup, Test_NoRamRegionIntialized) {
 }
 
 TEST(MfltStorageTestGroup, Test_CrashTracking) {
-  const size_t num_loops = 3;
-  for (size_t i = 0; i < num_loops; i++) {
-    memfault_reboot_tracking_mark_reset_imminent(kMfltRebootReason_UnknownError, NULL);
-    memfault_reboot_tracking_mark_reset_imminent(kMfltRebootReason_UserReset, NULL);
-  }
+  // Mark next bootup as created by an error
+  memfault_reboot_tracking_mark_reset_imminent(kMfltRebootReason_UnknownError, NULL);
 
-  // shouldn't impact the crash count
-  memfault_reboot_tracking_clear_reset_info();
+  // Crash count should not be incremented yet
+  size_t crash_count = memfault_reboot_tracking_get_crash_count();
+  LONGS_EQUAL(0, crash_count);
 
   // mcu reset reason being appended without a pre-existing reset reason should bump crash count
   sResetBootupInfo bootup_info = {
@@ -91,8 +90,13 @@ TEST(MfltStorageTestGroup, Test_CrashTracking) {
   };
   memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &bootup_info);
 
-  size_t crash_count = memfault_reboot_tracking_get_crash_count();
-  LONGS_EQUAL(num_loops + 1, crash_count);
+  crash_count = memfault_reboot_tracking_get_crash_count();
+  LONGS_EQUAL(1, crash_count);
+
+  // Resetting reboot tracking should not alter crash count
+  memfault_reboot_tracking_clear_reset_info();
+  crash_count = memfault_reboot_tracking_get_crash_count();
+  LONGS_EQUAL(1, crash_count);
 
   memfault_reboot_tracking_reset_crash_count();
   crash_count = memfault_reboot_tracking_get_crash_count();
@@ -105,6 +109,9 @@ TEST(MfltStorageTestGroup, Test_RebootSequence) {
   // 1. Watchdog Reboot - Confirm we get a Watchdog and then no more info
   memfault_reboot_tracking_mark_reset_imminent(kMfltRebootReason_SoftwareWatchdog, NULL);
   memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, NULL);
+
+  crash_count = memfault_reboot_tracking_get_crash_count();
+  LONGS_EQUAL(1, crash_count);
 
   sMfltResetReasonInfo info;
   bool info_available = memfault_reboot_tracking_read_reset_info(&info);
@@ -119,6 +126,9 @@ TEST(MfltStorageTestGroup, Test_RebootSequence) {
     .reset_reason = kMfltRebootReason_Assert,
   };
   memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &bootup_info);
+
+  crash_count = memfault_reboot_tracking_get_crash_count();
+  LONGS_EQUAL(2, crash_count);
   info_available = memfault_reboot_tracking_read_reset_info(&info);
   CHECK(info_available);
   LONGS_EQUAL(bootup_info.reset_reason, info.reason);
@@ -145,8 +155,130 @@ TEST(MfltStorageTestGroup, Test_RebootSequence) {
   info_available = memfault_reboot_tracking_read_reset_info(&info);
   CHECK(info_available);
   LONGS_EQUAL(kMfltRebootReason_Unknown, info.reason);
+  memfault_reboot_tracking_clear_reset_info();
 
-  // The POR and Firmware Update should not have been counted as a crash
+  // 5. Expected reboot due to firmware update, boot with an expected reboot reason
+  bootup_info.reset_reason = kMfltRebootReason_SoftwareReset;
+  memfault_reboot_tracking_mark_reset_imminent(kMfltRebootReason_FirmwareUpdate, NULL);
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &bootup_info);
+
+  info_available = memfault_reboot_tracking_read_reset_info(&info);
+  CHECK(info_available);
+  LONGS_EQUAL(kMfltRebootReason_FirmwareUpdate, info.reason);
+  memfault_reboot_tracking_clear_reset_info();
+
+  // 6. Boot with an expected reboot reason
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &bootup_info);
+  info_available = memfault_reboot_tracking_read_reset_info(&info);
+  CHECK(info_available);
+  LONGS_EQUAL(kMfltRebootReason_SoftwareReset, info.reason);
+
+  // Scenarios 5 and 6 should not count as unexpected crashes
   crash_count = memfault_reboot_tracking_get_crash_count();
-  LONGS_EQUAL(3, crash_count);
+  LONGS_EQUAL(4, crash_count);
+}
+
+TEST(MfltStorageTestGroup, Test_GetRebootReason) {
+  sMfltRebootReason reboot_reason = {
+    .reboot_reg_reason = kMfltRebootReason_UnknownError,
+  };
+  int rc = 0;
+
+  // Test that reason is invalid if arg is NULL
+  rc = memfault_reboot_tracking_get_reboot_reason(NULL);
+
+  LONGS_EQUAL(-1, rc);
+
+  // Test that reason is invalid before reboot tracking has booted
+  rc = memfault_reboot_tracking_get_reboot_reason(&reboot_reason);
+
+  LONGS_EQUAL(-1, rc);
+  LONGS_EQUAL((long)kMfltRebootReason_UnknownError, (long)reboot_reason.reboot_reg_reason);
+
+  // Test booting with null
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, NULL);
+
+  rc = memfault_reboot_tracking_get_reboot_reason(&reboot_reason);
+
+  LONGS_EQUAL(0, rc);
+  LONGS_EQUAL((long)kMfltRebootReason_Unknown, (long)reboot_reason.reboot_reg_reason);
+  LONGS_EQUAL((long)kMfltRebootReason_Unknown, (long)reboot_reason.prior_stored_reason);
+
+  // Test booting with a non-error reason, check that reboot reason is not cleared by
+  // memfault_reboot_tracking_clear_reset_info
+  memfault_reboot_tracking_clear_reset_info();
+  memfault_reboot_tracking_clear_reboot_reason();
+
+  sResetBootupInfo info = {
+    .reset_reason = kMfltRebootReason_SoftwareReset,
+  };
+
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &info);
+  rc = memfault_reboot_tracking_get_reboot_reason(&reboot_reason);
+
+  LONGS_EQUAL(rc, 0);
+  LONGS_EQUAL((long)kMfltRebootReason_SoftwareReset, (long)reboot_reason.reboot_reg_reason);
+  LONGS_EQUAL((long)kMfltRebootReason_SoftwareReset, (long)reboot_reason.prior_stored_reason);
+
+  memfault_reboot_tracking_clear_reset_info();
+  LONGS_EQUAL(rc, 0);
+  LONGS_EQUAL((long)kMfltRebootReason_SoftwareReset, (long)reboot_reason.reboot_reg_reason);
+  LONGS_EQUAL((long)kMfltRebootReason_SoftwareReset, (long)reboot_reason.prior_stored_reason);
+
+  // Simulate a crash, first boot with a crash. This will store the crash reason in reboot tracking
+  // and reboot reason memory. Check values
+  // Next clear reboot reason memory, and boot again with a different reason. Check values
+  info.reset_reason = kMfltRebootReason_Assert;
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &info);
+  rc = memfault_reboot_tracking_get_reboot_reason(&reboot_reason);
+
+  LONGS_EQUAL(rc, 0);
+  LONGS_EQUAL((long)kMfltRebootReason_Assert, (long)reboot_reason.reboot_reg_reason);
+  LONGS_EQUAL((long)kMfltRebootReason_Assert, (long)reboot_reason.prior_stored_reason);
+
+  memfault_reboot_tracking_clear_reboot_reason();
+  info.reset_reason = kMfltRebootReason_SoftwareReset;
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &info);
+  rc = memfault_reboot_tracking_get_reboot_reason(&reboot_reason);
+
+  LONGS_EQUAL(rc, 0);
+  LONGS_EQUAL((long)kMfltRebootReason_SoftwareReset, (long)reboot_reason.reboot_reg_reason);
+  LONGS_EQUAL((long)kMfltRebootReason_Assert, (long)reboot_reason.prior_stored_reason);
+}
+
+TEST(MfltStorageTestGroup, Test_GetUnexpectedRebootOccurred) {
+  bool unexpected_reboot = false;
+  int rc = 0;
+  sResetBootupInfo info = {.reset_reason = kMfltRebootReason_SoftwareReset};
+
+  // Test invalid if arg is NULL
+  rc = memfault_reboot_tracking_get_unexpected_reboot_occurred(NULL);
+  LONGS_EQUAL(-1, rc);
+
+  // Test invalid before reboot tracking has booted
+  rc = memfault_reboot_tracking_get_unexpected_reboot_occurred(&unexpected_reboot);
+  LONGS_EQUAL(-1, rc);
+
+  // Test booting with an expected reason
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &info);
+  rc = memfault_reboot_tracking_get_unexpected_reboot_occurred(&unexpected_reboot);
+  LONGS_EQUAL(0, rc);
+  CHECK_FALSE(unexpected_reboot);
+  memfault_reboot_tracking_clear_reset_info();
+
+  // Test booting with an unexpected reason
+  info.reset_reason = kMfltRebootReason_UnknownError;
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &info);
+  rc = memfault_reboot_tracking_get_unexpected_reboot_occurred(&unexpected_reboot);
+  LONGS_EQUAL(0, rc);
+  CHECK(unexpected_reboot);
+  memfault_reboot_tracking_clear_reset_info();
+
+  // Test booting with an unexpected reason marked, but expected reason at boot
+  info.reset_reason = kMfltRebootReason_PowerOnReset;
+  memfault_reboot_tracking_mark_reset_imminent(kMfltRebootReason_Assert, NULL);
+  memfault_reboot_tracking_boot(s_mflt_reboot_tracking_region, &info);
+  rc = memfault_reboot_tracking_get_unexpected_reboot_occurred(&unexpected_reboot);
+  LONGS_EQUAL(0, rc);
+  CHECK(unexpected_reboot);
 }